backend/requirements.txt
django==4.1.0
drf-spectacular==0.26.2
drf-social-oauth2==2.1.2
djangorestframework==3.14.0
django-oauth-toolkit==2.1.0
django-cors-headers==4.0.0
social-auth-app-django==5.2.0
django-filter==22.1
